Time: 5 ms (32.41%), Space: 6.2 MB (13.12%) - LeetHub

+++ 0029-divide-two-integers/0029-divide-two-integers.cpp
@@ -4,7 +4,7 @@ public:
     if(dividend==INT_MIN && divisor==-1) return INT_MAX;
     if(dividend==INT_MIN && divisor==1) return INT_MIN;
 
-      long long int  res = dividend/divisor;
+      long int  res = dividend/divisor;
         return res;
     }
 };
